+++ src/services/api/gifts/cartService.ts
@@ -1,7 +1,15 @@
 import axios from "axios";
 import { BASE_URL } from "../../../configuration/config";
 
-export async function addItem(token: string, item: string, value: string) {
+export async function addItem(
+  token: string | null,
+  item: string,
+  value: string
+) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.post(
     `${BASE_URL}/cart/add`,
     { item, value },
@@ -15,7 +23,11 @@ export async function addItem(token: string, item: string, value: string) {
   return res.data;
 }
 
-export async function getItems(token: string) {
+export async function getItems(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(`${BASE_URL}/cart/items`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -25,7 +37,11 @@ export async function getItems(token: string) {
   return res.data.items;
 }
 
-export async function getCartTotal(token: string) {
+export async function getCartTotal(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(`${BASE_URL}/cart/total`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -35,7 +51,11 @@ export async function getCartTotal(token: string) {
   return res.data;
 }
 
-export async function clearCart(token: string) {
+export async function clearCart(token: string | null) {
+  if (!token) {
+    return;
+  }
+  
   const res = await axios.delete(`${BASE_URL}/cart/clear`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -45,7 +65,11 @@ export async function clearCart(token: string) {
   return res.data;
 }
 
-export async function deleteItem(token: string, itemId: string) {
+export async function deleteItem(token: string | null, itemId: string) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.delete(`${BASE_URL}/cart/item/delete/${itemId}`, {
     headers: {
       Authorization: `Bearer ${token}`,
