+++ src/services/api/portal/sweetMessageService.ts
@@ -1,7 +1,11 @@
 import { BASE_URL } from "../../../configuration/config";
 import axios from "axios";
 
-export async function sendSweetMessage(token: string, message: string) {
+export async function sendSweetMessage(token: string | null, message: string) {
+  if (!token) {
+    return;
+  }
+  
   const res = await axios.post(
     `${BASE_URL}/sweet-messages/send-sweet-message`,
     { message },
@@ -15,7 +19,11 @@ export async function sendSweetMessage(token: string, message: string) {
   return res.data;
 }
 
-export async function getLastUnseenSweetMessage(token: string) {
+export async function getLastUnseenSweetMessage(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/sweet-messages/get-sweet-message/last-unseen`,
     {
@@ -28,7 +36,11 @@ export async function getLastUnseenSweetMessage(token: string) {
   return res.data;
 }
 
-export async function getSentSweetMessages(token: string) {
+export async function getSentSweetMessages(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/sweet-messages/get-sweet-messages/sent`,
     {
@@ -39,7 +51,11 @@ export async function getSentSweetMessages(token: string) {
   return res.data;
 }
 
-export async function getReceivedSweetMessages(token: string) {
+export async function getReceivedSweetMessages(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/sweet-messages/get-sweet-messages/received`,
     {
@@ -52,7 +68,11 @@ export async function getReceivedSweetMessages(token: string) {
   return res.data;
 }
 
-export async function viewSweetMessage(token: string, sweetId: string) {
+export async function viewSweetMessage(token: string | null, sweetId: string) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/sweet-messages/view-sweet-message/${sweetId}`,
     {
@@ -65,7 +85,11 @@ export async function viewSweetMessage(token: string, sweetId: string) {
   return res.data;
 }
 
-export async function deleteSweetMessage(token: string, sweetId: string) {
+export async function deleteSweetMessage(token: string | null, sweetId: string) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.delete(
     `${BASE_URL}/sweet-messages/delete-sweet-message/${sweetId}`,
     {
