+++ src/services/api/profiles/messageStorageService.ts
@@ -2,10 +2,14 @@ import axios from "axios";
 import { BASE_URL } from "../../../configuration/config";
 
 export async function storeMessage(
-  token: string,
+  token: string | null,
   title: string,
   message: string
 ) {
+  if (!token) {
+    return;
+  }
+  
   const res = await axios.post(
     `${BASE_URL}/messages/store`,
     { title, message },
@@ -50,11 +54,15 @@ export async function getReceivedMessages(token: string) {
 }
 
 export async function updateMessage(
-  token: string,
+  token: string | null,
   messageId: string,
   title: string,
   message: string
 ) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.put(
     `${BASE_URL}/messages/update/${messageId}`,
     {
@@ -71,7 +79,11 @@ export async function updateMessage(
   return res.data;
 }
 
-export async function deleteMessage(token: string, messageId: string) {
+export async function deleteMessage(token: string | null, messageId: string) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.delete(
     `${BASE_URL}/messages/delete/${messageId}`,
     {
